# PERF: Rolling.cov/corr